Refactored DiscoveredHost location, moved to internal/hosts
